moved some thing around to make it easier in the future
