+++ src/app/admin/page.tsx
@@ -2,7 +2,7 @@
 "use client";
 
 import React from "react";
-import ProjectManager from "../components/project-manager/ProjectManager";
+import ProjectManager from "./ProjectManager";
 
 export default function Page({ params }: { params: { title: string } }) {
   // Query the project based on the title
